refactor: move masking
